Fix CI

+++ src/save.c
@@ -1311,10 +1311,10 @@ read_f32(
 	if (stream_buf_pos + 4 > stream_buf_alloc_size)
 		return false;
 
-	*(uint32_t *)ret = (stream_buf[stream_buf_pos + 0] |
-			    (stream_buf[stream_buf_pos + 1] << 8) |
-			    (stream_buf[stream_buf_pos + 2] << 16) |
-			    (stream_buf[stream_buf_pos + 3] << 24));
+	*(uint32_t *)ret = ((uint32_t)stream_buf[stream_buf_pos + 0] |
+			    ((uint32_t)stream_buf[stream_buf_pos + 1] << 8) |
+			    ((uint32_t)stream_buf[stream_buf_pos + 2] << 16) |
+			    ((uint32_t)stream_buf[stream_buf_pos + 3] << 24));
 
 	stream_buf_pos += 4;
